# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: The Ministry of Herbology API
  version: 1.10.0
  description: |
    Contract for the Ministry of Herbology. Frozen — only the maintainers
    changes this file, via a recorded decision. Every path declares the parts of the project that
    implements it with `x-parts of the project`; `tests/contract/` enforces that the
    running API matches this document.

    Units: millimetres, °C, litres, lux. Timestamps are RFC 3339 UTC.
servers:
  - url: /api/v1
...
paths:
...
  /specimens:
    get:
      tags: [inventory]
      summary: List and search specimens
      operationId: listSpecimens
      parameters:
        - { name: q,           in: query, schema: { type: string } }
        - { name: location_id, in: query, schema: { type: string, format: uuid } }
        - { name: outdoor,     in: query, schema: { type: boolean } }
        - { name: status,      in: query, schema: { $ref: '#/components/schemas/SpecimenStatus' } }
        - { name: toxic_to_pets, in: query, schema: { type: boolean } }
        - { name: limit,  in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: A page of specimens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpecimenPage' }
    post:
      tags: [inventory]
      summary: Add a plant. Enrichment is queued and runs in the background.
      operationId: createSpecimen
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SpecimenCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Specimen' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /specimens/{specimen_id}:
    parameters:
      - { name: specimen_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [inventory]
      operationId: getSpecimen
      responses:
        '200':
          description: One specimen
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Specimen' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [inventory]
      operationId: updateSpecimen
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SpecimenUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Specimen' }
    delete:
      tags: [inventory]
      summary: Archive a specimen (soft delete)
      operationId: archiveSpecimen
      responses:
        '204': { description: Archived }
  /specimens/{specimen_id}/photos:
    parameters:
      - { name: specimen_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [inventory]
      operationId: listPhotos
      parameters:
        - name: order
          in: query
          required: false
          schema: { type: string, enum: [newest, oldest], default: newest }
          description: >-
            A guarantee since 1.7.0, not a description: `newest`
            is `taken_at` descending and is the default; `oldest` is ascending,
            which is how a growth log is read. Before 1.7.0 the order was prose
            and every client sorted for itself.
      responses:
        '200':
          description: Growth log photos, in the requested order (newest first by default)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Photo' }
    post:
      tags: [inventory]
      operationId: uploadPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:    { type: string, format: binary }
                caption: { type: string }
                taken_at:
                  type: string
                  format: date-time
                  description: When the photograph was taken; defaults to now. Declared in 1.8.0.
                is_primary:
                  type: boolean
                  default: false
                  description: >-
                    The Register's portrait for this plant. Exclusive per specimen:
                    setting it on one photograph clears it on the others. 1.8.0.
                member_id:
                  type: string
                  format: uuid
                  description: Who took it, recorded as `photo.taken_by`. Optional, never required. 1.8.0.
      responses:
        '201':
          description: Uploaded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Photo' }
  /specimens/{specimen_id}/photos/{photo_id}/image:
    parameters:
      - { name: specimen_id, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: photo_id,    in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [inventory]
      summary: A photograph's bytes
      operationId: getPhotoImage
      description: >-
        What `Photo.url` points at; with `variant=thumb`, what `Photo.thumb_url`
        points at. Declared in 1.8.0 for the third time on the
        reasoning the design gave `getMapLayerImage` and the design gave
        `getPlateImage`: the growth log fetches it for every photograph, so it
        has a client contract. The inventory API served it and declared it
        provisionally, as the maps module and the journal did.

        Served only under its own specimen's path: the same `photo_id` under
        another plant is a 404, because a photograph is a fact about one plant. The media type is the one the stored file's magic bytes
        identify. An upload has no generated thumbnail in 1.x, so
        `variant=thumb` on one serves the original.
      parameters:
        - name: variant
          in: query
          required: false
          schema: { type: string, enum: [thumb] }
      responses:
        '200':
          description: The image
          content:
            image/png:  { schema: { type: string, format: binary } }
            image/jpeg: { schema: { type: string, format: binary } }
        '404':
          description: >-
            No such photograph on this plant, or its file is missing from the
            volume — what a restore that took the database and not the photo
            volume looks like. Nothing is served in its place.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: No photo directory is configured. Names `MOH_PHOTOS_IMAGE_DIR`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /specimens/{specimen_id}/log:
    parameters:
      - { name: specimen_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [inventory]
      operationId: listLogEntries
      parameters:
        - name: order
          in: query
          required: false
          schema: { type: string, enum: [newest, oldest], default: newest }
          description: Same contract as `listPhotos`'s `order`, over `occurred_at`.
      responses:
        '200':
          description: Log entries, in the requested order (newest first by default)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LogEntry' }
    post:
      tags: [inventory]
      operationId: createLogEntry
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LogEntryCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LogEntry' }
  /locations:
    get:
      tags: [inventory]
      operationId: listLocations
      parameters:
        - { name: site_id, in: query, schema: { type: string, format: uuid } }
        - { name: outdoor, in: query, schema: { type: boolean } }
      responses:
        '200':
          description: The site → area → zone tree, flattened
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Location' }
    post:
      tags: [inventory]
      operationId: createLocation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LocationCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Location' }
  /locations/{location_id}:
    parameters:
      - { name: location_id, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [inventory]
      operationId: updateLocation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LocationCreate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Location' }
...
  /species:
    get:
      tags: [botany]
      operationId: listSpecies
      parameters:
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200':
          description: Known species
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Species' }
  /species/{species_id}:
    parameters:
      - { name: species_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [botany]
      operationId: getSpecies
      responses:
        '200':
          description: Species with its cited care profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Species' }
  /species/{species_id}/care-values:
    parameters:
      - { name: species_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [botany]
      summary: Every care field with its source and confidence
      operationId: listCareValues
      responses:
        '200':
          description: Cited care values
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CareValue' }
    put:
      tags: [botany]
      summary: User override of a care value
      operationId: overrideCareValue
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CareValueOverride' }
      responses:
        '200':
          description: Stored override
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CareValue' }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: moh_session
    # the stack's own services, GET only. Nginx deletes this header
    # from every inbound request, so it is unusable from outside the stack.
    service:
      type: apiKey
      in: header
      name: X-Moh-Service-Token
    # the Lunette's device token, valid on /hub/lunette only.
    lunette:
      type: http
      scheme: bearer
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
        code:   { type: string }
...
    SpecimenStatus:
      type: string
      enum: [thriving, struggling, dormant, overwintering, lost, given_away, archived]
...
    Specimen:
      type: object
      required: [id, is_outdoor, status, display_name]
      properties:
        id:              { type: string, format: uuid }
        display_name:    { type: string, description: 'nickname, else common name, else accepted name' }
        nickname:        { type: [string, 'null'] }
        cultivar:        { type: [string, 'null'] }
        species:         { $ref: '#/components/schemas/SpeciesBrief' }
        is_group:        { type: boolean }
        count:           { type: integer }
        location:        { $ref: '#/components/schemas/Location' }
        map_layer_id:    { type: [string, 'null'], format: uuid }
        pin_px:          { $ref: '#/components/schemas/PixelPoint' }
        is_outdoor:      { type: boolean }
        in_container:    { type: boolean }
        container_litres: { type: [number, 'null'] }
        soil_note:       { type: [string, 'null'] }
        acquired_on:     { type: [string, 'null'], format: date }
        provenance:      { type: [string, 'null'] }
        status:          { $ref: '#/components/schemas/SpecimenStatus' }
        primary_photo_url: { type: [string, 'null'] }
        next_task:
          # Nothing scheduled is a normal state, and a common one.
          anyOf: [{ $ref: '#/components/schemas/TaskBrief' }, { type: 'null' }]
        created_at:      { type: string, format: date-time }
...
    SpecimenCreate:
      type: object
      required: [name]
      properties:
        name:         { type: string, description: Typed name; resolved to a species }
        species_id:   { type: string, format: uuid, description: Skip resolution }
        image_key:    { type: string, description: Photo for Pl@ntNet identification }
        nickname:     { type: string }
        cultivar:     { type: string }
        location_id:  { type: string, format: uuid }
        in_container: { type: boolean, default: true }
        container_litres: { type: number }
        count:        { type: integer, default: 1 }
        acquired_on:  { type: string, format: date }
        taxon:
          description: >-
            The match the person picked from `POST /taxon/resolve`, exactly as
            returned. When given without `species_id`, the species is found by
            gbif_key, powo_id or accepted name, or created from it, in the same
            transaction; it is an identification, never a care fact. 1.10.0.
          allOf: [{ $ref: '#/components/schemas/TaxonCandidate' }]
...
    SpecimenPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Specimen' }
        next_cursor: { type: [string, 'null'] }
        total:       { type: integer }
...
security:
  - session: []